Backend/requirements.txt
fastapi
requests
uvicorn
python-jose